+++ src/global/global.service.ts
@@ -1,5 +1,5 @@
 import { HttpService } from '@nestjs/axios';
-import { BadRequestException, ForbiddenException, HttpException, Injectable, UnauthorizedException} from '@nestjs/common';
+import { BadRequestException, ForbiddenException, Injectable, UnauthorizedException} from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { InjectModel } from '@nestjs/mongoose';
 import {  AxiosError, AxiosRequestConfig } from 'axios';
@@ -7,7 +7,7 @@ import { Model } from 'mongoose';
 import { firstValueFrom, lastValueFrom,  } from 'rxjs';
 import { configuredHttpsAgent } from 'src/main';
 import { User } from 'src/schemas/user.schema';
-import { PathTree } from './types/types';
+import { PathTree, sbbAuthTree, sbbFintechTree } from 'libs/sbbApiEndpoints';
 
 @Injectable()
 export class GlobalService {
@@ -18,59 +18,15 @@ export class GlobalService {
         private readonly httpService: HttpService,
     ){}
 
-    private readonly sbbAuthTree: PathTree = {
-        base: {
-            url: this.configService.get('SB_ID_BASE_URL'),
-            children:[
-                {
-                    auth: {
-                        url: '/ic/sso/api/v2/oauth',
-                        children: [
-                            {
-                                authorize: {
-                                    url: '/authorize'
-                                }
-                            },
-                            {
-                                token: {
-                                    url:'/token'
-                                }
-                            },
-                            {
-                                clientInfo: {
-                                    url: '/user-info'
-                                }
-                            }
-                        ]
-                    },
-
-                },
-            ]
-        }
-    }
-
-    private readonly sbbFintechTree: PathTree = {
-        base: {
-            url: this.configService.get('SB_FINTECH_BASE_URL'),
-            children: [
-                {
-                    clientInfo: {
-                        url: '/v1/client-info' 
-                    }
-                },
-            ]
-        }
-    }
-
     getSbbUrl(str: string){
-        return this.getUrl(str, this.sbbAuthTree)
+        return this.getUrl(str, sbbAuthTree)
     }
 
     getFintechUrl(str: string){
-        return this.getUrl(str, this.sbbFintechTree)
+        return this.getUrl(str, sbbFintechTree)
     }
 
-    private getUrl(str: string, tree: PathTree = this.sbbAuthTree){
+    private getUrl(str: string, tree: PathTree){
         if(!str || !tree.base.url) return ''
         const points = str.split('.')
         const resultArray = [tree.base.url,]
@@ -147,8 +103,8 @@ export class GlobalService {
             }
         }
 
-        let responseData: Record<string, any> | undefined
-        let responseError: HttpException
+        try {
+            let responseData: Record<string, any> | undefined
         const user = await this.userModel.findOne({sbbAccessToken: accessToken});
         if(!user) throw new UnauthorizedException();
         const refreshToken = user.sbbRefreshToken
@@ -159,7 +115,8 @@ export class GlobalService {
         })
         .catch(async (error: AxiosError) => {
             if(!error?.response?.status) {
-                responseError =  new BadRequestException;
+                console.log('ONE 162');
+                throw  new BadRequestException;
             }else 
             if([401, 403].includes(error.response.status)){
                 await lastValueFrom(this.httpService.post(
@@ -189,22 +146,20 @@ export class GlobalService {
                     responseData = retryRes.data
                 })
                 .catch(() => {
-                  responseError =  new BadRequestException()
+                  throw  new BadRequestException()
                 })
             })
             .catch(() => {
-                responseError = new ForbiddenException()
+                throw new ForbiddenException()
             })
         
         } else {
-            responseError =  new BadRequestException()
+            throw  new BadRequestException()
         }
     })
-
-        if(responseError) {
-            throw responseError
-        } else {
-            return responseData
+        return responseData
+        } catch (error) {
+            throw error
         }
    
     }
