--- src/global/types/types.ts
@@ -1,6 +1,0 @@
-export type PathTree = {
-    [index: string]: {
-        url:string;
-        children?: PathTree[]
-    }
-}
